requirements.txt
cookiecutter >=1.7
gitpython
ruamel.yaml
